STYLE: rename file - delete last s
